Correction type

+++ src/main/java/cat/itacademy/barcelonactiva/Gispert/Judith/s04/t02/n02/Domain/Fruit.java
@@ -31,7 +31,7 @@ public class Fruit {
     public void setName (String name){
         this.name = name;
     }
-    public void setQuantityKg(int quantityKg){
+    public void setQuantityKg(float quantityKg){
         this.quantityKg = quantityKg;
     }
